third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/fetch-expected.txt

This is a testharness.js-based test.
PASS Same-host fetches are strictly same-site
PASS Subdomain fetches are strictly same-site
FAIL Cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.8806451537963156" but got (undefined) undefined
PASS Same-host redirecting to same-host fetches are strictly same-site
PASS Subdomain redirecting to same-host fetches are strictly same-site
PASS Cross-site redirecting to same-host fetches are strictly same-site
PASS Same-host redirecting to subdomain fetches are strictly same-site
PASS Subdomain redirecting to subdomain fetches are strictly same-site
PASS Cross-site redirecting to subdomain fetches are strictly same-site
FAIL Same-host redirecting to cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.42183907966731393" but got (undefined) undefined
FAIL Subdomain redirecting to cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.31949606557889476" but got (undefined) undefined
FAIL Cross-site redirecting to cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.1214731079967919" but got (undefined) undefined
Harness: the test ran to completion.
